change from timer to async for haptic feedback
